fix(latent-sft): 4 个连携 bug, 重命名 config.py 避开 import 冲突
复查 HLR 推理 + eval 全链时发现:

B1 (critical, 训练侧回归): train.py:215 LatentReasoner 没转 bf16, 只 to(device).
  主模型走 DeepSpeed bf16 后 hidden 是 bf16, 喂给 fp32 LR weight 做 Linear 会
  dtype mismatch. smoke_test_hlr.py:332 加了 bf16 转换, train.py 没加 → 第一次
  正式 train_hlr 才会暴露. 修复: latent_reasoner.to(accelerator.device).to(bfloat16).

B2 (推理一致性): inference.py 缺训练侧 "<think>\n" 防御逻辑.
  HLRDataset._build_sample (data_utils.py) 在 chat_template 后做了
  `if not prompt_text.rstrip().endswith("<think>"): prompt_text += "<think>\n"`.
  R1-Distill chat_template 不自动加 <think>, 推理时如果没补会让模型见到
  OOD 前缀分布. 修复: inference.py.generate 加同样的补 think 逻辑.

B3 (指标语义): evaluate.py hlr_summary.latent_savings_token_pct 是 HLR 自比
  (latent_steps × (ratio-1) / total_equiv), 不是"相对 baseline 节省"的真正含义,
  容易误导. 删除. 真正的节省指标走 eval_hlr_compare.py 跟 baseline 算.

B4 (critical, import 冲突): UniCOP-Reason/evaluate.py 顶部 `from config import config`
  会把 sys.modules["config"] 固定为 UniCOP-Reason/config.py. 后续走 backend=hlr
  分支 sys.path.insert Latent-SFT/ 再 `from config import HLRConfig` 不会重新
  load, 直接命中 cached UniCOP-Reason/config → ImportError (没有 HLRConfig).
  修复: 重命名 Latent-SFT/config.py → hlr_config.py, 同步改 train.py /
  inference.py / smoke_test_hlr.py / submit_train_hlr.sh 的 import.

+++ Latent-SFT/train.py
@@ -23,7 +23,7 @@ from torch.optim import AdamW
 from torch.utils.data import DataLoader
 from transformers import AutoModelForCausalLM, AutoTokenizer, get_cosine_schedule_with_warmup
 
-from config import HLRConfig
+from hlr_config import HLRConfig
 from data_utils import HLRDataset, collate_hlr
 from model import build_latent_reasoner_from_main, compute_hlr_loss
 
@@ -212,7 +212,10 @@ def train_hlr(cfg: HLRConfig):
     model, main_optimizer, dataloader = accelerator.prepare(
         model, main_optimizer, dataloader
     )
-    latent_reasoner = latent_reasoner.to(accelerator.device)
+    # LR 必须显式转 bf16: 主模型走 DeepSpeed bf16, hidden 是 bf16;
+    # LR 没进 accelerator.prepare, 默认 fp32 → 与 bf16 hidden 做 Linear 会 dtype mismatch.
+    # smoke_test_hlr.py:332 已经这么做, train.py 之前漏了 (回归 bug).
+    latent_reasoner = latent_reasoner.to(accelerator.device).to(torch.bfloat16)
 
     num_training_steps = math.ceil(
         len(dataloader) / cfg.gradient_accumulation_steps
